+++ zarrs/src/array/codec/array_to_bytes/vlen_v2/vlen_v2_codec.rs
@@ -61,7 +61,7 @@ impl ArrayCodecTraits for VlenV2Codec {
 
 #[cfg_attr(feature = "async", async_trait::async_trait)]
 impl ArrayToBytesCodecTraits for VlenV2Codec {
-    fn dynamic(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits> {
+    fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits> {
         self as Arc<dyn ArrayToBytesCodecTraits>
     }
 
@@ -154,7 +154,7 @@ impl ArrayToBytesCodecTraits for VlenV2Codec {
         ))
     }
 
-    fn compute_encoded_size(
+    fn encoded_representation(
         &self,
         decoded_representation: &ChunkRepresentation,
     ) -> Result<BytesRepresentation, CodecError> {